refactor: moved the alerts inside ProductFormViewController in a helper extension
